fix: integer truncation bug
This commit fixes a bug where the padded output array would have
incorrect dimensions due to doing the dimensional math on integers
instead of floats like the other branch.

This also refactors a bit to make this mistake harder to make and adds a
test that would have caught this bug.